Fix all compile errors
Most were trivial. The exceptions were

To<T> was added internally to Jint, which would cause us to diverge from the package. Fortunately, it did not depend on anything private, so I could write the code outside the package. Later, I should encapsulate in either an extension method or a utility function.

The second one is the handling of the delegateconversion, which again munged the Jint source code. I encapsulated most of the functionality into the constructor of a new OrchidTypeConverter class